refactor(proof): rename zk-backend support module to utils
