+++ crates/proof/zk/backend/src/succinct/dry_run.rs
@@ -5,7 +5,7 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-use super::support::get_range_elf_embedded;
+use super::utils::get_range_elf_embedded;
 use async_trait::async_trait;
 use base_proof_zk_host::{ZkProver, ZkProverError, ZkSessionState};
 use base_proof_zk_utils::client::DEFAULT_INTERMEDIATE_ROOT_INTERVAL;
