+++ crates/proof/zk/backend/src/succinct/utils/cache.rs
