+++ crates/proof/zk/backend/src/succinct/utils/cluster.rs
