+++ crates/proof/zk/backend/src/succinct/utils/contract.rs
