+++ crates/proof/zk/backend/src/succinct/utils/elf.rs
