+++ crates/proof/zk/backend/src/succinct/utils/mod.rs
