+++ crates/proof/zk/backend/src/succinct/utils/stats.rs
