+++ crates/proof/zk/backend/src/succinct/utils/stdin.rs
